cloudsec_backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
boto3
psycopg2-binary
gunicorn
pydantic
